# Conflicts:
#	hi_core/hi_core/HiseEventBuffer.cpp

+++ hi_modules/midi_processor/mps/MidiDelay.h
@@ -52,7 +52,7 @@ public:
 
 	void processHiseEvent(HiseEvent &m) override
 	{
-		m.setTimeStamp((uint16)delayTime);
+		m.setTimeStamp(delayTime);
 	};
